replace load_fst2 / free_fst2 by load_abstract_fst2 / free_abstract_Fst2 (except in LocatePattern.cpp) in AbstractFst2Load.h and AbstractFst2Load.cpp
The new load function can store an information about the need of free object
This can (and will be) be used later to implement a system of persistent FST2 object.

Note : On locatePattern.cpp, we call load_abstract_fst2 with free_info parameters as NULL, because 
we request an FST2 object which will be modified and free after by free_fst2.