introducing command line option only-verify-arguments
-V/--only-verify-arguments helps to check if the arguments
passed to a Unitex tool are syntactically correct and
enough to run it. If the test is successful a return value
of SUCCESS_RETURN_CODE (normally 0) is passed to the caller.
If not, a return value of USAGE_ERROR_CODE (64) is returned.
The tests are only performed in a syntax level, they will
not check e.g. if a filename passed as argument exists