[bin exp enhance] Enable outputs for subgraphs
Users can now have outputs on the boxes defined as subgraphs.
The output will be ignored if the first box inside the subgraph has a
loop or a dictionary variable in case of morphological mode. In other
cases, the output will be added to the first box of the subgraph. If
the box already has an output then the subgraph’s output will be
prepended