[major fix] LocateTfst honors the user's output policy on debug mode
Prior to this commit, for grammars compiled in debug mode,
LocateTfst produces always outputs in merge mode without follow
the user's output policy.

For any selected policy, this commit resolves the issue allowing
to produce the same outputs as those generated by a graph compiled
without the debug mode.

Notice that these changes are only related to graphs compiled in
debug mode to be applied with LocateTfst.