[major fix] Locate allows to debug graphs with dictionary-entry variables
Dictionary-entry variables in Locate (User's Manual, section 6.44) are
used to set variables to information stored in morphological-mode
dictionaries.

Prior to this commit, Locate was unable to debug a graph which use
dictionary-entry variables.

According to the Variable error policy, different types of errors are
produced:

- "Ignore variable errors": no error is shown but outputs involving the
use of dictionary-entry variables are missing.

- "Exit on variable error": the next error is thrown: "Output error:
undefined variable $foo$" and no matches are produced.

- "Backtrack on variable error": Locate loops infinitely.

This commit resolves the issue allowing to debug graphs with
dictionary-entry variables.

Notice that these changes are only related to graphs compiled in debug
mode to be applied with Locate.

Grammars which are not compiled on debug mode are not affected by this
bug. Grammars compiled in debug mode to be applied with LocateTfst are
affected with a related bug which is not fixed by this commit.