[major fix] LocateTfst allows to debug graphs with dictionary-entry variables
Dictionary-entry variables in LocateTfst (User's Manual, section 6.44)
are used to set variables to information contained in the lexical tags
of the text automaton.

Prior to this commit, LocateTfst was unable to debug a graph which use
dictionary-entry variables.

According to the Variable error policy, different types of errors are
produced:

- "Ignore variable errors": no error is shown but the outputs of boxes
containing dictionary-entry variables are replaced by the literal name
of the dictionary-entry variable. i.e. if $:foo$ is the variable then
foo will be the output.

- "Exit on variable error": the next error is thrown: "Output error:
missing closing $ after $" and no matches are produced.

- "Backtrack on variable error": LocateTfst produces not match.

This commit resolves the issue allowing to debug graphs with
dictionary-entry variables.

Notice that these changes are only related to graphs compiled in debug
mode to be applied with LocateTfst.

Grammars which are not compiled on debug mode are not affected by this
bug. Grammars compiled in debug mode to be applied with Locate are
affected with a related bug which is not fixed by this commit.