[break fix] Allow CasSys generalization graphs to match polylexical tags
This commit concerns the issue #59 and allows CasSys' generalization graphs
to match polylexical tags. The workaround used, as described in #59, is to
match polylexical tags using lexical masks, so that each unit recognized by
a generalization box, for example `foo-bar`, is now converted into a lexical
mask '<foo-bar>', i.e. surrounding the unit with the less-than and greater-than
signs.

See also #61