[tmp hotfix] Disable default VIRTOPTIMIZATION
This commit will disable memory virtual optimization features
which are introduced and enabled by default starting v4.0-alpha

A memory stack overflow is being raised while using e.g:

build_pattern(foo, bar, baz, p->al.prv_alloc_generic)

This issue needs further study to understand its origin.
That said, the problem could be not directly associated
with the introduction of the optmizations but rather to
the use of the generic locate allocator.

In the meantime, use VIRTOPTIMIZATION=yes to force the
compilation with this feature.