[fix Grf2Fst2] Allow to recognize the token @ in debug mode
During the compilation of grammar into a .fst2 file, double quotes
tokens are internally prefixed with '@' in order to indicate that
they don't tolerate case variations, e.g.:

"Fe" => @Fe

If a grammar uses the literal token @ and Grf2Fst2 is not compiling the
graph in debug mode the token @ is written as:

@ => %@

Here the prefix '%' indicates that a token can tolerate case variations.

Prior to this commit, if a grammar uses the literal token @ and Grf2Fst2 is
compiling the graph in debug mode (using the flag `--debug` ) the tag @
is written as:

@ => @

As result, the tag '@' will be interpreted as the strict-case prefix '@' and
the literal token @ will never be recognized.

This commit resolves this issue allowing to debug graphs containing the literal
token @, this is done by ensuring that the tag is written as:

@ => %@

Note that grammars which are not compiled on debug mode are not affected
by this bug.